Only display collective extra action in landing page actions
The actions depend on currentCollective, which is only available in
PageAction context, not in the collectives list.

Fixes: #475
